+++ cms_form/models/widgets/__init__.py
@@ -1,0 +1,10 @@
+from . import widget_mixin
+from . import widget_text
+from . import widget_hidden
+from . import widget_numeric
+from . import widget_boolean
+from . import widget_date
+from . import widget_selection
+from . import widget_many2one
+from . import widget_x2many
+from . import widget_binary
